+++ src/Web3Connector/MagicWeb3Connector.js
@@ -4,15 +4,9 @@ import AbstractWeb3Connector from './AbstractWeb3Connector';
 
 export default class MagicWeb3Connector extends AbstractWeb3Connector {
   type = 'MagicLink';
-  async activate({ email, apiKey, network }) {
+  async activate({ email, apiKey, network, newSession } = {}) {
     let magic = null;
     let ether = null;
-
-    try {
-      await this.deactivate();
-    } catch (error) {
-      // Do nothing
-    }
 
     if (!email) {
       throw new Error('"email" not provided, please provide Email');
@@ -43,6 +37,17 @@ export default class MagicWeb3Connector extends AbstractWeb3Connector {
       magic = new Magic(apiKey, {
         network: network,
       });
+
+      if (newSession) {
+        if (magic?.user) {
+          try {
+            await magic?.user?.logout();
+          } catch (error) {
+            // Do nothing
+          }
+        }
+      }
+
       ether = new ethers.providers.Web3Provider(magic.rpcProvider);
       await magic.auth.loginWithMagicLink({
         email: email,
@@ -79,5 +84,6 @@ export default class MagicWeb3Connector extends AbstractWeb3Connector {
     }
     this.account = null;
     this.chainId = null;
+    this.provider = null;
   };
 }